pyproject.toml
[tool.poetry]
name = "dnd-bot"
version = "1.0.0"
description = "Discord bot for ttrpg things, made for a close group of friends."
authors = ["Kirill <sir.blagish@gmail.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
beautifulsoup4 = "4.13.0b2"
discord-py = "2.4.0"
ply = "3.11"
requests = "2.32.3"
prettytable = "3.11.0"
loguru = ">=0.7.2,<0.8.0"
uvicorn = ">=0.30.6,<0.31.0"
fastapi = ">=0.112.1,<0.113.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
